+++ frontend/super-frontend/src/javascripts/super/application.ts
@@ -1,18 +1,15 @@
 import RailsUjs from "@rails/ujs";
-import { Application, Controller } from "stimulus";
+import { Application, Controller as StimulusController } from "stimulus";
 import ApplyTemplateController from "./apply_template_controller";
 import TogglePendingDestructionController from "./toggle_pending_destruction_controller";
 
 RailsUjs.start();
 
-const application = Application.start();
-application.register("apply-template", ApplyTemplateController);
-application.register(
+let StimulusApplication = Application.start();
+StimulusApplication.register("apply-template", ApplyTemplateController);
+StimulusApplication.register(
   "toggle-pending-destruction",
   TogglePendingDestructionController
 );
 
-export default {
-  StimulusApplication: application,
-  StimulusController: Controller,
-};
+export { StimulusApplication, StimulusController };